feat: cs-1905 Add button to PDF template Editor to Generate PDF
