--- apps/tenant-management-webapp/src/app/pages/admin/services/pdf/templates/previewEditor/styled-components.ts
@@ -1,243 +1,0 @@
-import styled, { createGlobalStyle } from 'styled-components';
-import { PreviewPortal } from '../previewPortal';
-export const NotificationTemplateEditorContainer = styled.div`
-  display: flex;
-  flex: auto;
-  padding-left: 3rem;
-  width: 100%;
-  height: 100vh;
-  overflow: hidden;
-  box-sizing: border-box;
-`;
-// Edit Template components
-export const TemplateEditorContainer = styled.div`
-  padding-right: 1rem;
-  flex: auto;
-  margin-top: 4rem;
-  overflow: hidden;
-  &:hover {
-    overflow: auto;
-  }
-
-  .reduce-margin {
-    margin-top: 5px;
-  }
-
-  @media (min-width: 1279px) {
-    .mobile {
-      display: none;
-    }
-  }
-
-  @media (max-width: 1280px) {
-    .desktop {
-      display: none;
-    }
-
-    .mobile > div {
-      padding: 2px 0 2px 3px;
-    }
-  }
-`;
-
-export const TemplateEditorContainerPdf = styled.div`
-  padding-right: 1rem;
-  flex: auto;
-  overflow: hidden;
-  &:hover {
-    overflow: auto;
-  }
-
-  .reduce-margin {
-    margin-top: 5px;
-  }
-
-  @media (min-width: 1279px) {
-    .mobile {
-      display: none;
-    }
-  }
-
-  .goa-form-item {
-    margin-bottom: 0rem !important;
-  }
-
-  .title {
-    font-size: var(--fs-base);
-    font-weight: var(--fw-bold);
-    padding-bottom: 12px;
-    padding-top: 8px;
-  }
-
-  @media (max-width: 1280px) {
-    .desktop {
-      display: none;
-    }
-
-    .mobile > div {
-      padding: 2px 0 2px 3px;
-    }
-  }
-`;
-export const Modal = styled.div`
-  display: block;
-  position: fixed;
-  left: 0;
-  z-index: 10000;
-  width: 100%;
-`;
-
-export const BodyGlobalStyles = createGlobalStyle<{ hideOverflow: boolean }>`
-  body {
-    overflow:  ${(props) => (props.hideOverflow ? `hidden` : `auto`)};
-  }
-`;
-export const ModalContent = styled.div`
-  background: white;
-`;
-export const PreviewTemplateContainer = styled.div`
-  width: 612px;
-  margin-left: 2rem;
-  padding-top: 1rem;
-  padding-left: 2rem;
-  background-color: #c3c3c3;
-  overflow: hidden;
-  &:hover {
-    overflow: auto;
-  }
-`;
-export const MonacoDiv = styled.div`
-  display: flex;
-  border: 1px solid var(--color-gray-700);
-  border-radius: 3px;
-  padding: 0.15rem 0.15rem;
-`;
-export const MonacoDivBody = styled.div`
-  display: flex;
-  border: 1px solid var(--color-gray-700);
-  border-radius: 3px;
-  padding: 0.15rem 0.15rem;
-  height: calc(92vh - 376px - 7rem);
-  min-height: 2rem;
-`;
-export const MonacoDivHeader = styled.div`
-  display: flex;
-  border: 1px solid var(--color-gray-700);
-  border-radius: 3px;
-  padding: 0.15rem 0.15rem;
-  height: calc(46vh - 210px - 3.5rem);
-  min-height: 2rem;
-`;
-export const MonacoDivFooter = styled.div`
-  display: flex;
-  border: 1px solid var(--color-gray-700);
-  border-radius: 3px;
-  padding: 0.15rem 0.15rem;
-  height: calc(46vh - 210px - 3.5rem);
-  min-height: 2rem;
-`;
-export const EditTemplateActions = styled.div`
-  display: flex;
-  justify-content: right;
-  gap: 1rem;
-`;
-
-// preview template components
-export const PreviewContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: 85%;
-  margin-right: 2rem;
-`;
-
-export const SubjectPreview = styled.div`
-  background-color: white;
-  padding-left: 1rem;
-`;
-
-export const BodyPreview = styled(PreviewPortal)`
-  background-color: white;
-  overflow: hidden;
-  &:hover {
-    overflow: auto;
-  }
-  flex-grow: 1;
-  margin-bottom: 1rem;
-`;
-
-export const PdfEditorLabelWrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-  .badge {
-    margin: 0px 0px 0px 3px;
-  }
-`;
-
-export const PdfConfigFormWrapper = styled.div`
-  .goa-form-item {
-    margin-bottom: 0.2rem !important;
-  }
-  padding-left: 3px;
-  padding-top: 1.5rem;
-`;
-
-export const PdfEditActionLayout = styled.div`
-  position fixed;
-  bottom: 2rem;
-`;
-
-export const PdfEditActions = styled.div`
-  display: flex;
-  gap: 0.5rem;
-`;
-export const EditorStyles = styled.div`
-  border: 1px solid #666;
-  border-radius: 3px;
-`;
-
-export const SpinnerSpace = styled.div`
-  margin: 10px 9px 10px 14px;
-  float: right;
-`;
-
-export const GenerateButtonPadding = styled.div`
-  margin: 0 0 0 14px;
-`;
-
-export const PaddingRight = styled.div`
-  margin-right: 12px;
-  padding-top: 1rem;
-`;
-
-export const FlexRow = styled.div`
-  display: flex;
-  flex-direction: row;
-`;
-export const GeneratorStyling = styled.div`
-  .extra-padding {
-    margin: 20px 0 0 0;
-  }
-
-  .topBottomPadding {
-    padding: 15px 0;
-  }
-
-  .row-flex {
-    display: flex;
-    flex-direction: row;
-  }
-
-  .indicator {
-    background: #f3f3f3;
-    min-width: 160px;
-  }
-
-  .event-stream {
-    flex: 3;
-    fontsize: 12px;
-  }
-
-  .button-margin {
-    margin: 0 0 20px 0;
-  }
-`;
