--- apps/tenant-management-webapp/src/app/pages/admin/services/pdf/testGenerate/styled-components.ts
@@ -1,48 +1,0 @@
-import styled from 'styled-components';
-export const FileTableStyles = styled.div`
-  .flex-horizontal {
-    display: flex;
-    flex-direction: row;
-  }
-
-  .flex {
-    flex: 1;
-  }
-
-  .mt-1 {
-    margin-top: 2px;
-  }
-
-  .mt-2 {
-    margin-top: 4px;
-  }
-`;
-
-export const EditorStyles = styled.div`
-  border: 1px solid #666;
-  border-radius: 3px;
-`;
-
-export const SpinnerPadding = styled.div`
-  margin: 0 0 0 5px;
-  float: right;
-`;
-
-export const SpinnerSpace = styled.div`
-  margin: 10px 9px 10px 14px;
-  float: right;
-`;
-
-export const GenerateButtonPadding = styled.div`
-  margin: 0 0 0 14px;
-`;
-
-export const PaddingRight = styled.div`
-  margin-right: 12px;
-`;
-
-export const FlexRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  padding-top: 1rem;
-`;
